Require delays to be unsigned.
This causes integer values being passed to lead to "overloaded call
ambiguity" error, and that's the intention, because using signed's
(which are C's default type for integer literals) may lead to strange
ambiguities, for example 65535's type is "long" (32 bit).

+++ include/timer_msp430.hpp
@@ -40,13 +40,13 @@ public:
         delay_since(start, cycles);
     }
 
-    static void delay(int32_t cycles)
+    static void delay(uint32_t cycles)
     {
         width start = value();
         while (true) {
             width end = value();
             cycles -= end - start;
-            if (cycles < 0)
+            if ((int32_t)cycles < 0)
                 break;
             start = end;
         }
